requirements.txt
autogen-agentchat[llm,gemini]~=0.2
flaml[automl]==2.3.2
colorama~=0.4.6